moves mem-related functions to utils
